Update README.md
A chat bot that replies to users' emoji reactions which can belong to a happy,a sad or an angry group. The chat bot remembers the previous state of the user, and the bot's reply reflects the change in the user's mood. An SQL database stores different data of each user.